Add files via upload
In train1.py I used HOG of SVM in place of our implemented hog method because are the same.
I added the HOM method to the code for calculating the HOG & HOM discriptor.
I added also the HARD NEGATIVES MEANING method.
A hard negative is when you take that falsely detected patch, and explicitly create a negative example out of that patch, and add that negative to your training set. When you retrain your classifier, it should perform better with this extra knowledge, and not make as many false positives.